CfApp becomes BlueGreenApp, gets own space.
[#83972842]
