+++ lib/bookbinder/blue_green_app.rb
@@ -1,0 +1,27 @@
+module Bookbinder
+  class BlueGreenApp
+    def initialize(name)
+      @name = name.strip
+    end
+
+    def ==(other)
+      to_s == other.to_s
+    end
+
+    def to_s
+      name
+    end
+
+    def with_flipped_name
+      if name.include?('green')
+        BlueGreenApp.new(name.sub('green', 'blue'))
+      else
+        BlueGreenApp.new(name.sub('blue', 'green'))
+      end
+    end
+
+    private
+
+    attr_reader :name
+  end
+end
